consolidate styles
