+++ Client/Screens/Clue.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { Button, Image, ScrollView, StyleSheet, Text, View } from 'react-native'
+import { Button, Image, ScrollView, Text, View } from 'react-native'
+import { styles } from '../../assets/styles';
 
 const Clue = (props) => {
   return (
@@ -14,7 +15,7 @@ const Clue = (props) => {
             source={require('../../assets/1234.jpg')}
             style={styles.winImage}
           />
-          <View style={styles.startButtonContainer}>
+          <View style={styles.greenButtonContainer}>
             <Button
               title="Continue"
               onPress={() => { props.navigation.navigate('Map') }}
@@ -26,65 +27,5 @@ const Clue = (props) => {
     </ScrollView>
   )
 }
-
-const styles = StyleSheet.create({
-  clueContainer: {
-    display: 'flex',
-    backgroundColor: 'green',
-    flexDirection: 'column',
-    height: '100%',
-    width: '100%',
-    padding: 50,
-  },
-  headerContainer: {
-    backgroundColor: 'yellow',
-    width: '100%',
-    borderRadius: 5,
-    // borderWidth: 3,
-    // borderColor: 'black',
-    padding: 20,
-  },
-  header: {
-    fontSize: 30,
-    color: 'black',
-    textAlignVertical: 'center',
-    textAlign: 'center',
-  },
-  winImage: {
-    marginBottom: 30,
-    width: 270
-  },
-  bodyContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 30,
-  },
-  helloWorldTextStyle: {
-    fontFamily: 'Arial',
-    fontSize: 30,
-    color: '#ffffff',
-    textAlignVertical: 'center',
-    textAlign: 'center',
-  },
-  start: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: 'green'
-  },
-  startButtonContainer: {
-    backgroundColor: 'goldenrod',
-    padding: 10,
-    borderRadius: 5,
-    // borderWidth: 3,
-    // borderColor: 'black',
-  },
-  startButton: {
-    color: 'black'
-  },
-});
 
 export default Clue
